src/tools/information_provider.py
"""
Information Tool Provider for the tool integration framework.

This module implements the Information Tool Provider which provides tools for
searching and retrieving information from external sources.
"""

import logging
from typing import Dict, List, Any, Optional, Union, Callable

from tools.models import ToolCategory, ParameterType, ToolRegistry
from tools.provider import ToolProvider


class InformationToolProvider(ToolProvider):
    """
    Provides tools for searching and retrieving information from external sources.
    """
    
    def __init__(self, tool_registry: ToolRegistry):
        """
        Initialize the InformationToolProvider.
        
        Args:
            tool_registry: The tool registry to register tools with.
        """
        super().__init__(tool_registry)
    
    def register_tools(self) -> None:
        """Register information tools with the tool registry."""
        # Web Search
        self.register_tool(
            self.create_tool(
                name="info_search_web",
                description="Search web pages using search engine.",
                category=ToolCategory.INFORMATION,
                function=self.info_search_web,
                parameters=[
                    self.create_parameter(
                        name="query",
                        description="Search query in Google search style, using 3-5 keywords.",
                        type=ParameterType.STRING,
                        required=True
                    ),
                    self.create_parameter(
                        name="date_range",
                        description="(Optional) Time range filter for search results. Defaults to \"all\" (no time restriction). Use other options only when explicitly required by the task.",
                        type=ParameterType.STRING,
                        required=False,
                        default="all",
                        enum=["all", "past_hour", "past_day", "past_week", "past_month", "past_year"]
                    )
                ]
            )
        )
        
        # Image View
        self.register_tool(
            self.create_tool(
                name="image_view",
                description="View image content.",
                category=ToolCategory.INFORMATION,
                function=self.image_view,
                parameters=[
                    self.create_parameter(
                        name="image",
                        description="Absolute path of the image file to view",
                        type=ParameterType.STRING,
                        required=True
                    )
                ]
            )
        )
    
    # Tool implementations
    def info_search_web(self, query: str, date_range: str = "all") -> Dict[str, Any]:
        """
        Search the web for information.
        
        Args:
            query: The search query.
            date_range: Time range filter for search results.
            
        Returns:
            Dict[str, Any]: The search results.
        """
        try:
            # In a real implementation, this would perform a web search
            # For this prototype, we'll return simulated search results
            
            # Validate date range
            valid_ranges = ["all", "past_hour", "past_day", "past_week", "past_month", "past_year"]
            if date_range not in valid_ranges:
                return {
                    "success": False,
                    "error": f"Invalid date range: {date_range}. Must be one of: {', '.join(valid_ranges)}"
                }
            
            # Simulate search results
            results = [
                {
                    "title": "Example Search Result 1",
                    "url": "https://example.com/result1",
                    "snippet": "This is an example search result snippet that matches the query.",
                    "date": "2025-03-01",
                    "position": 1
                },
                {
                    "title": "Example Search Result 2",
                    "url": "https://example.com/result2",
                    "snippet": "Another example search result with relevant information.",
                    "date": "2025-02-15",
                    "position": 2
                },
                {
                    "title": "Example Search Result 3",
                    "url": "https://example.com/result3",
                    "snippet": "A third example search result with additional context.",
                    "date": "2025-01-20",
                    "position": 3
                }
            ]
            
            return {
                "success": True,
                "query": query,
                "date_range": date_range,
                "results": results,
                "result_count": len(results)
            }
        
        except Exception as e:
            self.logger.error(f"Error searching web: {str(e)}")
            return {
                "success": False,
                "error": str(e)
            }
    
    def image_view(self, image: str) -> Dict[str, Any]:
        """
        View an image file.
        
        Args:
            image: Path to the image file.
            
        Returns:
            Dict[str, Any]: The image metadata and content reference.
        """
        try:
            # Check if file exists and is an image
            import os
            import imghdr
            
            if not os.path.isfile(image):
                return {
                    "success": False,
                    "error": f"Image file not found: {image}"
                }
            
            # Check if it's a valid image file
            image_type = imghdr.what(image)
            if not image_type:
                return {
                    "success": False,
                    "error": f"Not a valid image file: {image}"
                }
            
            # Get image metadata
            file_size = os.path.getsize(image)
            
            # In a real implementation, this would process the image for viewing
            # For this prototype, we'll return simulated metadata
            
            return {
                "success": True,
                "image_path": image,
                "image_type": image_type,
                "file_size": file_size,
                "dimensions": "1920x1080",  # Simulated dimensions
                "content_reference": f"image://{image}"  # Reference to the image content
            }
        
        except Exception as e:
            self.logger.error(f"Error viewing image: {str(e)}")
            return {
                "success": False,
                "error": str(e)
            }
