requirements.txt
# Python dependencies for Personal AI Agent

# Core dependencies
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
typer>=0.9.0

# LLM and embedding dependencies
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2
langchain>=0.0.200
langchain-community>=0.0.10

# Vector database
qdrant-client>=1.5.0

# Tool integration dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
pillow>=10.0.0
python-multipart>=0.0.6

# Development dependencies
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
